introduced available_minimizers module
