+++ tests/unit_tests/Fitting/minimizers/test_minimizer_lmfit.py
@@ -3,11 +3,10 @@ import pytest
 from unittest.mock import MagicMock
 
 import easyscience.fitting.minimizers.minimizer_lmfit
-
+from easyscience.fitting import AvailableMinimizers
 from easyscience.fitting.minimizers.minimizer_lmfit import LMFit
 from easyscience.Objects.new_variable import Parameter
 from lmfit import Parameter as LMParameter
-from easyscience.Objects.ObjectClasses import BaseObj
 from easyscience.fitting.minimizers.utils import FitError
